RavenDB-17890 - Export DummyWebSocket into a class.
